Notify transition ready for player before taking snapshot
So SnapshotController and shell's onTransitionReady can run in parallel.
That reduces the latency before starting the animator.

Theoretically, the binder transmission time of transition info is
usually enough to complete the snapshot. However, even if hardcoding
hundreds millisecond to delay the snapshot until the transition is
finished, the snapshot is still captured successfully, so it should be
safe to reorder.

Bug: 279437990
Bug: 297493149
Test: TransitionTests#testTransientLaunch
      KeyboardVisibilityControlTest
Test: Disable animation. Launch/back between different tasks.
      The valid snapshot of the closed task can still be shown in recent.
Change-Id: Id3cc1cc47412c098c8c8e8d63a09367772527400