Consider transition change info for task snapshot
For example, when launching an activity in a new task, the activity
requests to change orientation at runtime before it is drawn.
The transition info will be like:
 type=OPEN {
  All changes: sb=(0, 0 - 1080, 2340) eb=(0, 0 - 2340, 1080) r=0->1
  opening task: mode=OPEN
  closing task: mode=TO_BACK
  display: mode=CHANGE
 }

Because the configuration of tasks will be updated with display,
the closing task also has the configuration in new orientation.
But the activity in the closing task still keeps the appearance
in original orientation, so the attributes of snapshot should be
consistent with it. Otherwise the task snapshot may show a rotated
and cropped activity.

Also remove "inFinishingTransition" condition in createSnapshot
because currently RecentsTransitionHandler will take snapshot
from ATMS#takeTaskSnapshot before the transition is finished.

Bug: 301670257
Bug: 279678674
Test: TaskSnapshotControllerTest
Change-Id: If030f175bc60804d8b1f90fc88b670789fee4026