Request transtion when PiP gets removed
I4c13ecc1b3c1dff15bd0d61aeafeacc1d7705460 added a call to
`requestCloseTransitionIfNeeded()` in
`removePinnedRootTaskInSurfaceTransaction` to ensure that a shell
transition runs when PiP gets removed, but the CL was reverted
due to a CTS regression.

Iec89dcdd0081eca1f26161aaead1662f0f661ab6 added the logic not to
introduce the regression but in the CL we forgot to handle the
original intention of calling `requestCloseTransitionIfNeeded()`,
which this CL takes care of.

One note is that, it turns out removing PiP is acutally not CLOSE
but TO_BACK given that it just involves windowing mode and z-order
changes, so with this CL, a TO_BACK transition is requested.

Bug: 303767495
Test: atest PinnedStackTests
Change-Id: I0821cf62ef54c7f41266b2e8bbc6927da9778483