Fix broken @link references in CompatModePackages
The javadoc for @ChangeId constants are generated differently to normal
docs into a compatchanges.html page. This page is kind of compiled
stand-alone, so can only resolve references to the public SDK.
CompatModePackages itself is not in the SDK, so the @link references
in this class fail to resolve.

Fix this by using manually defined links (<a>) instead. Note that
doclava generates headers with the id of the compat constant that we can
use for the href of our <a> tags.

Bug: 303570856
Bug: 303184203
Test: m offline-sdk-docs ; inspect compatchanges.html
Change-Id: I5785098554f65b377c8f862b6e3a50225703f742