Play SLEEP transitions if they contain changes
Since queued transitions aren't always cleanly chunked
the way we'd imagine them to be, we may end up with
actual events in the transition info if we happen to
be turning screens on/off/foldingat the same time as doing
things with the device.

In this case, the SLEEP transition would have come first
and the actual changes second, so we should assume these
are changes that need to be played.

Change-Id: I4bbaa8ad033aaadcb1c0b54889b7c81fb003b001
Test: atest ShellTransitionTests
Bug: 304713150