Do no use text sub for empty string in sim.
In sim screen, we often will not have the subscription info when we
factory reset. This often leads to "Enter simpin for """. In the case
that the display name is an empty string, we want to just revert back to
the default message.

Fixes: 293964211
Test: show sim screen with multiple sim cards, one of which does not
have a display name in the subscription info.

Change-Id: Ibd06987e84811edd1387889b052ea1153fdf0ce0