clobber wom: implement onKeyUp for PowerKeyRule
- Apply the callback for power key. This is helpful for devices that
wish to trigger power key press behavior without waiting for
the delay needed to check for consecutive presses.

- Added a new config, `config_shortPressEarlyOnPower` to guard Power key's
usage of this callback. If this config is enabled, the single press
behavior for power key will run in response to the first key up event,
instead of using the `onPress` callback.

Test: atest WmTests:SingleKeyGestureTests
BUG: 305856148
Change-Id: I9f69c0352ecd8ee495dcbc0d9d5967635b02f69d