Hide Private tabs when in quiet mode
Check hideInSettingsInQuietMode user property -
If this user property is set then we should check if the user is in
quiet mode or not - only when the user is not in quiet mode, should we
show this user in Settings.

This is for the Private Profile which sets this user property. Without
this change, the Private profile tab is always shown irrespective of the
quiet mode.

Also, check the SHOW_IN_SETTINGS_SEPARATE instead of
SHOW_IN_LAUNCHER_SEPARATE

Bug: 309935782
Test: manual
Change-Id: Icaef8d853932236246bac169ad83399fbef18bec