Fix janky auto-enter PiP animation from split activity
This happens to Activity Embedding split and system split screen.

When minimizing the split pair of activities with one activity auto
entering PiP. It used to keep the other split activity visible without
any animation until the entering PiP animation (bounds) is done, which
looks janky.

Ideally, the transition should be splitted so that the PiP change is
animated by the PipTransition while the rest should be picked up by
other handler. This is a short term fix to immediately hide the other
closing window.

Fix: 306988504
Test: manual testing: in three-button navigation, show AE split
      activities with auto-pip, and press home button.
Change-Id: Ib9425814d21eb419d8906f1c3580a7083c052fb3