Animate QS between Shade and QuickSettings
Use the transitionState to drive the expanding animation in the adapter.
Now that we are using movable content, we don't need to make sure to
attach and detach manually.

Quick fix is needed in QSAnimator so that after the animators are
updated (and all the views are reset), we set the progress to the last
progress. This is an issue in compose because the view is dettached and
reattached at the end of the animation.

Fixes: 312503265
Test: atest QSSceneAdapterImplTest QSSceneAdapterTest
Test: manual: enable flags and expand/collapse
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: I2e1c86eea01f9926797209224c5d10e2f6bee06d