Log BAL allow state if hardening would be enabled
Currently we only know the state of the BAL allow state that is actually
used, but to evaluate what will happen if the feature is enabled, we
also need the state we would get if the feature was enabled.

This only adds a field to the logs. No behavior change expected.

Bug: 296478951
Test: atest BackgroundActivityLaunchTest
Change-Id: I2585976bdb5248e7ec9c0ae1bb9ecad4b41c6dcd